+++ irs-api/src/main/java/org/eclipse/tractusx/irs/semanticshub/SemanticsHubCacheInitializer.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -43,7 +44,7 @@ class SemanticsHubCacheInitializer {
     private final List<String> defaultUrns;
 
     /* package */ SemanticsHubCacheInitializer(final SemanticsHubFacade semanticsHubFacade,
-            @Value("${semanticsHub.defaultUrns:}") final List<String> defaultUrns) {
+            @Value("${semanticshub.defaultUrns:}") final List<String> defaultUrns) {
         this.semanticsHubFacade = semanticsHubFacade;
         this.defaultUrns = defaultUrns;
     }
@@ -62,17 +63,30 @@ class SemanticsHubCacheInitializer {
                 log.error("Initialization of semantic hub cache failed for URN '{}'", urn, ex);
             }
         });
+        try {
+            semanticsHubFacade.getAllAspectModels();
+        } catch (SchemaNotFoundException e) {
+            log.error("Initialization of semantic model cache failed.", e);
+        }
     }
 
     /**
      * Cleaning up Semantics Hub cache after scheduled time, and reinitializing it once again.
      */
-    @Scheduled(cron = "${semanticsHub.cleanup.scheduler}")
+    @Scheduled(cron = "${semanticshub.cleanup.scheduler}")
     /* package */ void reinitializeAllCacheInterval() {
         log.debug("Reinitializing Semantics Hub Cache with new values.");
 
-        semanticsHubFacade.evictAllCacheValues();
-        initializeCacheValues();
+        defaultUrns.stream().findFirst().ifPresentOrElse(urn -> {
+            try {
+                semanticsHubFacade.getModelJsonSchema(urn);
+                log.info("Could retrieve schema. Reinitializing cache values");
+                semanticsHubFacade.evictAllCacheValues();
+                initializeCacheValues();
+            } catch (SchemaNotFoundException e) {
+                log.error("Error while retrieving semantic models for cache. Reusing existing cached values", e);
+            }
+        }, () -> log.warn("Semantic models could not be retrieved for cache. Reusing existing cached value"));
     }
 
 }