+++ irs-api/src/test/java/org/eclipse/tractusx/irs/IrsApplicationTests.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -21,14 +22,18 @@
  ********************************************************************************/
 package org.eclipse.tractusx.irs;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.io.File;
+import java.io.InputStream;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import org.awaitility.Awaitility;
 import org.eclipse.tractusx.irs.aaswrapper.job.AASTransferProcess;
 import org.eclipse.tractusx.irs.aaswrapper.job.ItemDataRequest;
 import org.eclipse.tractusx.irs.component.JobParameter;
@@ -40,7 +45,6 @@ import org.eclipse.tractusx.irs.connector.job.JobOrchestrator;
 import org.eclipse.tractusx.irs.connector.job.JobStore;
 import org.eclipse.tractusx.irs.connector.job.ResponseStatus;
 import org.eclipse.tractusx.irs.persistence.BlobPersistence;
-import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -72,10 +76,14 @@ class IrsApplicationTests {
 
     @Test
     void generatedOpenApiMatchesContract() throws Exception {
-        final String generatedYaml = this.restTemplate.getForObject("http://localhost:" + port + "/api/api-docs.yaml",
-                String.class);
-        final String fixedYaml = Files.readString(new File("../api/irs-v1.0.yaml").toPath(), UTF_8);
-        assertThat(generatedYaml).isEqualToIgnoringWhitespace(fixedYaml);
+        final String generatedYaml = this.restTemplate.getForObject("http://localhost:" + port + "/api/api-docs.yaml", String.class);
+        final InputStream fixedYaml = Files.newInputStream(Path.of("../api/irs-v1.0.yaml"));
+
+        final ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+        final Map<String, Object> fixedYamlMap = mapper.readValue(fixedYaml, Map.class);
+        final Map<String, Object> generatedYamlMap = mapper.readValue(generatedYaml, Map.class);
+
+        assertThat(fixedYamlMap).isEqualTo(generatedYamlMap);
     }
 
     @Test