+++ irs-api/src/test/java/org/eclipse/tractusx/irs/connector/job/MultiTransferJobTest.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -51,9 +52,9 @@ class MultiTransferJobTest {
 
     @Test
     void getJobData_Immutable() {
-        final List<AspectType> aspectTypes = jobParameter().getAspects();
-
-        assertThatExceptionOfType(UnsupportedOperationException.class).isThrownBy(() -> aspectTypes.add(AspectType.ID_CONVERSION));
+        final List<String> aspectTypes = jobParameter().getAspects();
+        final String conversionId = AspectType.ID_CONVERSION.toString();
+        assertThatExceptionOfType(UnsupportedOperationException.class).isThrownBy(() -> aspectTypes.add(conversionId));
     }
 
     @Test